@@ -2,12 +2,14 @@ package ru.oliverhd.weather;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -29,7 +31,6 @@ public class CitiesFragment extends Fragment implements Constants {
 
         String[] data = getResources().getStringArray(R.array.cities);
         initRecyclerView(data, view);
-
 //        initList(view);
     }
 
@@ -48,7 +49,6 @@ public class CitiesFragment extends Fragment implements Constants {
         if (isExistWeatherDetail) {
             showWeatherDetail(currentParcel);
         }
-
     }
 
     @Override
@@ -80,7 +80,6 @@ public class CitiesFragment extends Fragment implements Constants {
                             .addToBackStack(null)
                             .commit();
                 }
-
             }
         });
     }
@@ -121,7 +120,7 @@ public class CitiesFragment extends Fragment implements Constants {
                         .replace(R.id.weather_info, detail)
                         .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
                         .commit();
-            }else {
+            } else {
                 getFragmentManager()
                         .beginTransaction()
                         .replace(R.id.weather_info, WeatherDetailFragment.create(parcel))
