@@ -1,7 +1,8 @@
 package ru.oliverhd.weather;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -32,7 +33,6 @@ public class DetailAdapter extends RecyclerView.Adapter<DetailAdapter.ViewHolder
     public int getItemCount() {
         return dataSource.length;
     }
-
 
     public class ViewHolder extends RecyclerView.ViewHolder {
         private TextView time;
