@@ -1,7 +1,9 @@
 package ru.oliverhd.weather;
 
 import android.content.res.Configuration;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import android.os.Bundle;
 
 public class MainActivity extends AppCompatActivity {
@@ -12,16 +14,10 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
-
             getSupportFragmentManager()
                     .beginTransaction()
                     .replace(R.id.fragment_container, new MainFragment())
                     .commit();
-
-//            getSupportFragmentManager()
-//                    .beginTransaction()
-//                    .replace(R.id.fragment_container, new CitiesFragment())
-//                    .commit();
         } else {
             getSupportFragmentManager()
                     .beginTransaction()
